ai/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-multipart
PyPDF2
pytesseract
Pillow
